+++ src/simulator/SimulatorGameOfLifeTest.h
@@ -1,0 +1,6 @@
+#pragma once
+#ifdef SIMULATOR
+class GfxRenderer;
+class MappedInputManager;
+bool verifySimulatorGameOfLifeRules(GfxRenderer& renderer, MappedInputManager& input);
+#endif